feat: add new blog post about being accepted to Buildspace S4
This commit adds a new blog post titled "I have been accepted to the Buildspace S4" to the _posts directory. The blog post provides information about Buildspace, an online bootcamp that teaches how to build a SaaS product. It also mentions the author's excitement about being accepted to the S4 cohort and their plans to focus on a new side project during the 6-week bootcamp.

The blog post also mentions the author's idea for a community website called DeskSpace, where people can share their desk setups. The author is unsure if the idea will be validated but sees potential in it. The tech stack for the project is mentioned as NextJS, Prisma, tRPC, TailwindCSS, and Supabase.

Link to the blog post: [_posts/2023-08-02-Buildspace.md](https://github.com/username/repo/blob/commit